+++ src/Grafana.OpenTelemetry.Base/Instrumentations/HostResource.cs
@@ -1,0 +1,24 @@
+//
+// Copyright Grafana Labs
+// SPDX-License-Identifier: Apache-2.0
+//
+
+#if !NETSTANDARD
+
+using OpenTelemetry.ResourceDetectors.Host;
+using OpenTelemetry.Resources;
+
+namespace Grafana.OpenTelemetry
+{
+    internal class HostResourceInitializer : InstrumentationInitializer
+    {
+        public override Instrumentation Id { get; } = Instrumentation.HostResource;
+
+        protected override ResourceBuilder InitializeResourceDetector(ResourceBuilder builder)
+        {
+            return builder.AddDetector(new HostDetector());
+        }
+    }
+}
+
+#endif