+++ src/Grafana.OpenTelemetry.Base/Instrumentations/InstrumentationInitializer.cs
@@ -6,6 +6,7 @@
 using System;
 using OpenTelemetry.Logs;
 using OpenTelemetry.Metrics;
+using OpenTelemetry.Resources;
 using OpenTelemetry.Trace;
 
 namespace Grafana.OpenTelemetry
@@ -33,7 +34,15 @@ namespace Grafana.OpenTelemetry
             new QuartzInitializer(),
             new SqlClientInitializer(),
             new StackExchangeRedisInitializer(),
-            new WcfInitializer()
+            new WcfInitializer(),
+            new AWSResourceInitializer(),
+            new AzureResourceInitializer(),
+            new ContainerResourceInitializer(),
+#if !NETSTANDARD
+            new HostResourceInitializer(),
+            new ProcessResourceInitializer(),
+            new ProcessRuntimeResourceInitializer()
+#endif
         };
 
         abstract public Instrumentation Id { get; }
@@ -43,6 +52,7 @@ namespace Grafana.OpenTelemetry
             try
             {
                 InitializeTracing(builder);
+                InitializeResource(builder);
 
                 GrafanaOpenTelemetryEventSource.Log.EnabledTracingInstrumentation(Id.ToString());
             }
@@ -57,6 +67,7 @@ namespace Grafana.OpenTelemetry
             try
             {
                 InitializeMetrics(builder);
+                InitializeResource(builder);
 
                 GrafanaOpenTelemetryEventSource.Log.EnabledMetricsInstrumentation(Id.ToString());
             }
@@ -66,10 +77,25 @@ namespace Grafana.OpenTelemetry
             }
         }
 
+        protected void InitializeResource(TracerProviderBuilder builder)
+        {
+            builder.ConfigureResource(resourceBuilder => InitializeResourceDetector(resourceBuilder));
+        }
+
+        protected void InitializeResource(MeterProviderBuilder builder)
+        {
+            builder.ConfigureResource(resourceBuilder => InitializeResourceDetector(resourceBuilder));
+        }
+
         protected virtual void InitializeTracing(TracerProviderBuilder builder)
         { }
 
         protected virtual void InitializeMetrics(MeterProviderBuilder builder)
         { }
+
+        protected virtual ResourceBuilder InitializeResourceDetector(ResourceBuilder builder)
+        {
+            return builder;
+        }
     }
 }