+++ src/Grafana.OpenTelemetry.Base/Instrumentations/ProcessResource.cs
@@ -1,0 +1,24 @@
+//
+// Copyright Grafana Labs
+// SPDX-License-Identifier: Apache-2.0
+//
+
+#if !NETSTANDARD
+
+using OpenTelemetry.ResourceDetectors.Process;
+using OpenTelemetry.Resources;
+
+namespace Grafana.OpenTelemetry
+{
+    internal class ProcessResourceInitializer : InstrumentationInitializer
+    {
+        public override Instrumentation Id { get; } = Instrumentation.ProcessResource;
+
+        protected override ResourceBuilder InitializeResourceDetector(ResourceBuilder builder)
+        {
+            return builder.AddDetector(new ProcessDetector());
+        }
+    }
+}
+
+#endif