+++ src/Grafana.OpenTelemetry.Base/Instrumentations/ProcessRuntimeResource.cs
@@ -1,0 +1,24 @@
+//
+// Copyright Grafana Labs
+// SPDX-License-Identifier: Apache-2.0
+//
+
+#if !NETSTANDARD
+
+using OpenTelemetry.ResourceDetectors.ProcessRuntime;
+using OpenTelemetry.Resources;
+
+namespace Grafana.OpenTelemetry
+{
+    internal class ProcessRuntimeResourceInitializer : InstrumentationInitializer
+    {
+        public override Instrumentation Id { get; } = Instrumentation.ProcessRuntimeResource;
+
+        protected override ResourceBuilder InitializeResourceDetector(ResourceBuilder builder)
+        {
+            return builder.AddDetector(new ProcessRuntimeDetector());
+        }
+    }
+}
+
+#endif